+++ lib/dartapi_core.dart
@@ -23,6 +23,7 @@ export 'src/core/health_controller.dart';
 export 'src/openapi/security_scheme.dart';
 export 'src/openapi/openapi_generator.dart';
 export 'src/openapi/docs_controller.dart';
+export 'src/openapi/query_param_spec.dart';
 
 export 'src/utils/utils.dart';
 export 'src/utils/multipart.dart';